Framework
An example of how to implement a discrete (n-dimensional) fields. These fields could then be generated using user input or affect each other based on a rule. (Could be poisson's eq,  maxwell, schrodinger (especially in this case). We could create arrays of fields to represent vector fields or tensor fields so this naturally generalizes to electrodynamics or general relativity.